+++ src/potent/display_modes.py
@@ -3,11 +3,13 @@ Some high-level structs that centrally control output verbosity levels
 """
 
 from dataclasses import dataclass
-from typing import Literal, Sequence, get_args
+from typing import Annotated, Literal, Sequence, get_args
 
-from cyclopts import Token
+from cyclopts import Parameter, Token
 
+# don't need this type if it's just for key names? Could be dict keys
 DisplayModeNames = Literal["quiet", "standard", "verbose"]
+display_mode_names = get_args(DisplayModeNames)
 
 
 @dataclass
@@ -23,7 +25,7 @@ class DisplayMode:
         mode = tokens[0].value
         if mode not in modes_by_name:
             raise ValueError(
-                f"{mode} is not a valid display mode. Options are: {get_args(DisplayModeNames)}"
+                f"{mode} is not a valid display mode. Options are: {display_mode_names}"
             )
 
         return modes_by_name[mode]
@@ -50,3 +52,15 @@ VerboseDisplayMode = DisplayMode(
 _modes = [QuietDisplayMode, StandardDisplayMode, VerboseDisplayMode]
 
 modes_by_name: dict[DisplayModeNames, DisplayMode] = {m.name: m for m in _modes}
+
+DisplayModeFlag = Annotated[
+    DisplayMode,
+    Parameter(
+        name="display-mode",
+        converter=DisplayMode.parse,
+        # https://github.com/BrianPugh/cyclopts/issues/886
+        help=f"Controls how the results are displayed. [choices:  {', '.join(display_mode_names)}] (default: standard)",
+        accepts_keys=False,
+        show_default=False,
+    ),
+]